kyc-copilot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
opencv-python-headless==4.10.0.84
Pillow==10.4.0
python-multipart==0.0.9
pydantic==2.8.2
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.2
python-dotenv==1.0.1
redis==5.0.8
boto3==1.35.0
